requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic>=2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
supabase==1.2.0
httpx>=0.24.0,<0.25.0
pytest==7.4.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
loguru==0.7.2
